Fix example to run with webpack 5
webpack serve should be used instead of webpack-dev-server starting
with webpack 5.
An explicit static directory ("/" in this case) needs to be specified
to ensure the index.html file is found.
The new webpack expects an explicit mode option.
Set it to "development".
Finally an extra leading slash is needed in output.publicPath to
ensure the compiled js file is found.

Amends 56539f6e042e7bc1f855380a639db777a4112d18